requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.4
django-cors-headers==3.12.0
django-phonenumber-field==6.1.0
django-phonenumbers==1.0.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
phonenumbers==8.12.46
Pillow==9.1.0
PyJWT==2.3.0
pytz==2022.1
shortuuid==1.0.9
sqlparse==0.4.2
